CURL

+++ ajax/EscolaJSONAjax.php
@@ -39,48 +39,61 @@ switch ($_REQUEST["acao"])
 
 		if ($_GET["idesc"])
 		{
-			$dompdf = new Dompdf();
-
-			$host = $_SERVER["HTTP_HOST"];
-			echo $host;
-			die();
-
-			$folder = ($_SERVER["HTTP_HOST"] == "187.73.149.26:8080" ? "" : "Hospital/";
-			$file = file_get_contents("http://{$host}/{$folder}pesquisa_pdf.php?idesc=".$_GET["idesc"]);
-
-			// Carrega o conteúdo da página
-			$dompdf->load_html($file);
-
-			// Define o tamanho da página para A4 e orientação para retrato
-			$dompdf->setPaper('A4', 'portrait');
-
-			// Gera a visualização do arquivo
-			$dompdf->render();
-
-			// Exporta o arquivo
-			$arquivo = $dompdf->output("arquivo.pdf", array("Attachment" => true));
-
-			// Gera um nome criptografado para o arquivo
-			$rand = rand(1,500);
-			$rand2 = rand(1,500);
-			$nomeCrip = md5("arquivo".$rand.$rand2);
-
-			if ($arquivo)
+			if (function_exists("curl_init"))
 			{
-				file_put_contents($path['arquivos'].$nomeCrip.'.pdf',$arquivo);
+				$dompdf = new Dompdf();
+				$host = $_SERVER["HTTP_HOST"];
+				$folder = ($_SERVER["HTTP_HOST"] == "187.73.149.26:8080" ? "" : "Hospital/");
+				$url = "http://{$host}/{$folder}pesquisa_pdf.php?idesc=".$_GET["idesc"];
 
-				$env = new EnvioDocumento();
-			    $env->setEnv_idEscola($_GET['idesc']);
-			    $env->setEnv_idRemetente(1);
-			    $env->setEnv_idDestinatario(4);
-			    $env->setEnv_url($nomeCrip.'.pdf');
-			    $env->setVisto(0);
+				// Iniciar o CURL
+				$fileReader = curl_init();
 
-				$envioDocumentoControler->insert($env);
+				// Configurar a leitura e o retorno da URL
+				curl_setopt($fileReader, CURLOPT_URL, $url);
+				curl_setopt($fileReader, CURLOPT_HEADER, 0);
+				curl_setopt($fileReader, CURLOPT_RETURNTRANSFER, 1);
+				curl_setopt($fileReader, CURLOPT_USERAGENT, 'Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US; rv:1.7.5) Gecko/20041107 Firefox/1.0');
 
-				echo json_encode(["status" => "Arquivo gerado com sucesso", "arquivo" => $nomeCrip.".pdf"]);
-			} else
-				throw new Exception("Ocorreu um erro ao gerar o arquivo pdf.");
+				// Executa a leitura da URL
+				$file = curl_exec($fileReader);
+
+				// Carrega o conteúdo da página
+				$dompdf->load_html($file);
+
+				// Define o tamanho da página para A4 e orientação para retrato
+				$dompdf->setPaper('A4', 'portrait');
+
+				// Gera a visualização do arquivo
+				$dompdf->render();
+
+				// Exporta o arquivo
+				$arquivo = $dompdf->output("arquivo.pdf", array("Attachment" => true));
+
+				// Gera um nome criptografado para o arquivo
+				$rand = rand(1,500);
+				$rand2 = rand(1,500);
+				$nomeCrip = md5("arquivo".$rand.$rand2);
+
+				if ($arquivo)
+				{
+					file_put_contents($path['arquivos'].$nomeCrip.'.pdf',$arquivo);
+
+					$env = new EnvioDocumento();
+				    $env->setEnv_idEscola($_GET['idesc']);
+				    $env->setEnv_idRemetente(1);
+				    $env->setEnv_idDestinatario(4);
+				    $env->setEnv_url($nomeCrip.'.pdf');
+				    $env->setVisto(0);
+
+					$envioDocumentoControler->insert($env);
+
+					echo json_encode(["status" => "Arquivo gerado com sucesso", "arquivo" => $nomeCrip.".pdf"]);
+				} else
+					throw new Exception("Ocorreu um erro ao gerar o arquivo pdf.");
+			}
+			else
+				throw new Exception("CURL não configurado.", 1);
 		}
 		else
 			throw new Exception("Parâmetro 'idesc' ausente.", 1);
